minor refactor FrameSimulatorPluginUDP
