--- frameSimulator/src/DummyUDPFrame.cpp
@@ -1,9 +1,0 @@
-#include "DummyUDPFrame.h"
-
-namespace FrameSimulator {
-
-    DummyUDPFrame::DummyUDPFrame(const int& frameNum) : frame_number(frameNum) {
-
-    }
-
-}
