+++ frameSimulator/src/UDPFrame.cpp
@@ -1,0 +1,9 @@
+#include "UDPFrame.h"
+
+namespace FrameSimulator {
+
+    UDPFrame::UDPFrame(const int& frameNum) : frame_number(frameNum) {
+
+    }
+
+}
