+++ test/integrationTest/CMakeLists.txt
@@ -1,3 +1,4 @@
 # Add subdirectories for integrationTest
 add_subdirectory(include)
 add_subdirectory(src)
+add_subdirectory(config)
