+++ test/integrationTest/src/HDF5FrameTest.cpp
@@ -49,8 +49,10 @@ namespace FrameSimulatorTest {
               std::string output_file = ptree.get<std::string>("Test.output_file");
               PropertyTreeUtility::expandEnvVars(output_file);
 
+	      std::string dataset_name = "/" + ptree.get<std::string>("Test.dataset");
+
               file_id = H5Fopen(output_file.c_str(), H5F_ACC_RDONLY, H5P_DEFAULT);
-              dataset = H5Dopen(file_id, "/data", H5P_DEFAULT);
+              dataset = H5Dopen(file_id, dataset_name.c_str(), H5P_DEFAULT);
             } else {
               throw std::runtime_error("HDF5FrameTest: json file not specified!");
             }
